+++ entities/server.js
@@ -3,13 +3,15 @@
  * 
  * @param  string name The name of the entity
  * @param  object options Passed to the find() method
- * @return object
  */
 Meteor.publish('entity', function(name, options) {
 	options = options ? options : {};
 	return orion.entities[name].collection.find(options);
 });
 
+/**
+ * Publication for the table on the admin.
+ */
 Meteor.publish('entityTabular', function(tableName, ids, fields) {
 	check(tableName, String);
 	check(ids, [String]);